+++ NEWS.md
@@ -1,6 +1,9 @@
-# snahelper 0.1.1
+# snahelper 0.2.0
 
 * added option to repel node labels
+* removed curvature for edges and added manual alpha tuning
+* switched from `geom_edge_arc` to `geom_edge_fan`
+* changed some default colors
 
 # snahelper 0.1.0